--- ci/go_test.sh
@@ -1,2 +1,0 @@
-go test ./
-go test ./util/iplocation
