--- common/const.go
@@ -1,5 +1,0 @@
-package common
-
-const (
-	WindowsCallInvalidError = "The operation completed successfully."
-)
