+++ common/constants.go
@@ -1,0 +1,8 @@
+package common
+
+const (
+	// WindowsCallInvalidError 定义了一个常量字符串，用于表示Windows系统调用出错时的无效错误信息。
+	// 这个错误信息实际上表明操作已经成功完成，但它可能被用作一个默认的错误消息，
+	// 在实际的错误信息未被正确设置或获取时使用。
+	WindowsCallInvalidError = "The operation completed successfully."
+)
