+++ go.mod
@@ -2,4 +2,4 @@ module github.com/skys-mission/gout
 
 go 1.18
 
-require golang.org/x/sys v0.20.0
+require golang.org/x/sys v0.27.0
