+++ go/os/win/privilegew/doc/manifest.xml
