+++ go/os/win/privilegew/privilegew_test.go
