+++ go/os/win/processw/example/rsrc_windows_amd64.syso
