--- util/iplocation/const.go
@@ -1,7 +1,0 @@
-package iplocation
-
-const (
-	baseURL = "https://api.iplocation.net"
-
-	getIPAddrPath = "/?cmd=get-ip"
-)
